Split custom hooks for applying single responsability
