+++ src/hooks/useGistDetails.tsx
@@ -1,0 +1,11 @@
+import { useEffect, useState } from 'react';
+import { PublicGist } from '~src/@types';
+
+export const useGistDetails = () => {
+  const [gistData, setGistData] = useState<PublicGist>();
+  const [isFetching, setIsFetching] = useState<boolean>(true);
+
+  useEffect(() => {}, []);
+
+  return { gistData, isFetching };
+};
